Move logging specs inside correct new describe class method block and refactor logger class slightly to avoid name conflict
